requirements.txt
pydantic==1.9.0
scikit-learn==1.0.2
boto3==1.20.34
python-dotenv==0.19.2